+++ adapters/Chartboost/ChartboostAdapter/GADMAdapterChartboostUtils.h
@@ -19,6 +19,7 @@
 #endif
 #import <Foundation/Foundation.h>
 #import <GoogleMobileAds/GoogleMobileAds.h>
+#import "GADMediationAdapterChartboost.h"
 
 #define SYSTEM_VERSION_LESS_THAN(v)                    \
   ([[[UIDevice currentDevice] systemVersion] compare:v \
@@ -51,6 +52,11 @@ NSString *_Nonnull GADMAdapterChartboostLocationFromAdConfiguration(
 
 /// Creates and returns a Chartboost mediation object.
 CHBMediation *_Nonnull GADMAdapterChartboostMediation(void);
+
+/// Returns an NSError with code |code| and with NSLocalizedDescriptionKey and
+/// NSLocalizedFailureReasonErrorKey values set to |description|.
+NSError *_Nonnull GADMAdapterChartboostErrorWithCodeAndDescription(
+    GADMAdapterChartboostErrorCode code, NSString *_Nonnull description);
 
 /// Returns the closest CHBBannerSize size from the requested GADAdSize.
 CHBBannerSize GADMAdapterChartboostBannerSizeFromAdSize(